requirements.txt
langchain
langchain-community
langchain-openai
langchain-huggingface
openai
transformers
sentence-transformers
chromadb
gradio
pandas
numpy
python-dotenv
